requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Faker==37.12.0
idna==3.11
iniconfig==2.3.0
loguru==0.7.3
packaging==25.0
pluggy==1.6.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
